[+] Added test for analyzing the data present within the exams search.
